refactor: remove xjs.Parse and xjs.Print
since xjs is a library for creating JavaScript dialects, not for
parsing or printing JavaScript. These applications had no useful
function for the end user.
