+++ xjs_test.go
@@ -4,22 +4,32 @@ import (
 	"fmt"
 
 	"github.com/xjslang/xjs"
+	"github.com/xjslang/xjs/js"
 )
 
 func Example_basic() {
 	input := `function hello() {
-	let x = 100
-	let y = 200
-}`
-	result, err := xjs.Parse([]byte(input))
+		let x = 100
+		let y = 200
+		}`
+
+	// code --> AST
+	sc := xjs.ScannerBuilder().Build([]byte(input))
+	p := xjs.ParserBuilder().Build(sc)
+	result, err := js.ParseProgram(p)
 	if err != nil {
 		panic(err)
 	}
-	out, err := xjs.Print(result)
+
+	// AST --> code
+	pr := xjs.PrinterBuilder().Build()
+	pr.Print(result)
+	out, err := pr.Output()
 	if err != nil {
 		panic(err)
 	}
 	fmt.Print(out)
+
 	// Output:
 	// function hello() {
 	//   let x = 100;
